refactor: EventTagRepository 경로 수정
- repository 패키지로 이동

#86
