+++ backend/emm-sale/src/main/java/com/emmsale/event/application/EventService.java
@@ -14,8 +14,8 @@ import com.emmsale.event.application.dto.EventResponse;
 import com.emmsale.event.domain.Event;
 import com.emmsale.event.domain.EventStatus;
 import com.emmsale.event.domain.EventTag;
-import com.emmsale.event.domain.EventTagRepository;
 import com.emmsale.event.domain.repository.EventRepository;
+import com.emmsale.event.domain.repository.EventTagRepository;
 import com.emmsale.event.exception.EventException;
 import com.emmsale.tag.domain.Tag;
 import com.emmsale.tag.domain.TagRepository;
