+++ backend/emm-sale/src/main/java/com/emmsale/event/domain/repository/EventTagRepository.java
@@ -1,5 +1,6 @@
-package com.emmsale.event.domain;
+package com.emmsale.event.domain.repository;
 
+import com.emmsale.event.domain.EventTag;
 import com.emmsale.tag.domain.Tag;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
