+++ backend/emm-sale/src/test/java/com/emmsale/event/application/EventServiceTest.java
@@ -18,8 +18,8 @@ import com.emmsale.event.application.dto.EventDetailResponse;
 import com.emmsale.event.application.dto.EventResponse;
 import com.emmsale.event.domain.Event;
 import com.emmsale.event.domain.EventTag;
-import com.emmsale.event.domain.EventTagRepository;
 import com.emmsale.event.domain.repository.EventRepository;
+import com.emmsale.event.domain.repository.EventTagRepository;
 import com.emmsale.event.exception.EventException;
 import com.emmsale.event.exception.EventExceptionType;
 import com.emmsale.helper.ServiceIntegrationTestHelper;
